requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
pytest==7.4.2
gunicorn==21.2.0
psycopg2-binary==2.9.7
Flask-Migrate==4.0.5 